added explicit int cast

+++ src/render.cu
@@ -58,7 +58,7 @@ int main(void) {
     npy::npy_data d = npy::read_npy<float>(npy_path);
     std::vector<float> npy_data = d.data;
     std::vector<unsigned long> npy_shape = d.shape;
-    int mb_shape[3] = {npy_shape[0], npy_shape[1], npy_shape[2]};
+    int mb_shape[3] = {(int)npy_shape[0], (int)npy_shape[1], (int)npy_shape[2]};
     int data_size = npy_data.size();
     float *p_data = npy_data.data();
     size_t bytes = data_size * sizeof(float);
